+++ lib/repositories/openai/openai_realtime_repository.dart
@@ -127,7 +127,6 @@ class OpenaiRealtimeRepository extends RealtimeModel {
 
   @override
   void commitUserAudio() {
-    _logger.debug('Committing user audio');
     var mapData = {
       "type": "input_audio_buffer.commit",
     };
@@ -137,7 +136,6 @@ class OpenaiRealtimeRepository extends RealtimeModel {
 
   @override
   void sendUserAudio(Uint8List audioData) {
-    _logger.debug('Sending user audio');
     var mapData = {
       "type": "input_audio_buffer.append",
       "audio": base64Encode(audioData),
@@ -208,7 +206,6 @@ class OpenaiRealtimeRepository extends RealtimeModel {
     s.stream.listen(
       (event) async {
         String rawData = event;
-        _logger.debug('Received event: $rawData');
         Map<String, dynamic> data = jsonDecode(rawData);
         String type = data['type'];
         await _processServerMessage(type, data);
@@ -252,10 +249,13 @@ class OpenaiRealtimeRepository extends RealtimeModel {
 
         for (var limit in rateLimits) {
           if (limit['name'] == 'requests') {
-            _onRemainingRequestsUpdated.add(limit['remaining']);
+            num amount = limit['remaining'];
+            _onRemainingRequestsUpdated.add(amount.toInt());
           } else if (limit['name'] == 'tokens') {
-            _onRemaingTimeUpdated
-                .add(Duration(seconds: limit['reset_seconds']));
+            num seconds = limit['reset_seconds'];
+            _onRemaingTimeUpdated.add(Duration(
+              seconds: seconds.toInt(),
+            ));
           }
         }
       },
@@ -327,8 +327,15 @@ class OpenaiRealtimeRepository extends RealtimeModel {
       _logger.warn('No handler found for type: $type');
       return;
     }
+    _logger.debug('Processing server message of type: $type');
 
-    await handler();
+    try {
+      await handler();
+    } on Exception catch (e) {
+      _onError.add(e);
+      _logger.error('Error while processing $type: $e');
+      _logger.error('Received data: $data');
+    }
   }
 
   @override